requirements.txt
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
fastapi==0.103.2
h11==0.14.0
idna==3.4
pydantic_core==2.10.1
sniffio==1.3.0
starlette==0.27.0
uvicorn==0.23.2
